+++ src/components/HURUmap/Tutorial/ToolTipItem/ToolTipIcon.js
